Rename syntax_highlighter to rust2html
