+++ homework/81621/rust2html/sample_inputs/simple.rs
